EVG-3357: migrate plugin handlers to service package
