+++ makefile
@@ -3,7 +3,7 @@ name := evergreen
 buildDir := bin
 nodeDir := public
 packages := $(name) agent operations cloud command db subprocess util plugin hostinit units
-packages += plugin-builtin-attach plugin-builtin-manifest plugin-builtin-buildbaron plugin-builtin-perfdash
+packages += plugin-builtin-attach plugin-builtin-manifest plugin-builtin-buildbaron
 packages += thirdparty alerts auth scheduler model validator service monitor repotracker
 packages += model-patch model-artifact model-host model-build model-event model-task model-user model-distro model-testresult model-version
 packages += model-grid rest-client rest-data rest-route rest-model migrations spawn model-trigger
