+++ model/task/task.go
@@ -2,7 +2,9 @@ package task
 
 import (
 	"fmt"
+	"regexp"
 	"sort"
+	"strings"
 	"time"
 
 	"github.com/evergreen-ci/evergreen"
@@ -39,6 +41,10 @@ const (
 
 var (
 	AgentHeartbeat = "heartbeat"
+
+	// A regex that matches either / or \ for splitting directory paths
+	// on either windows or linux paths.
+	eitherSlash *regexp.Regexp = regexp.MustCompile(`[/\\]`)
 )
 
 type Task struct {
@@ -1499,4 +1505,29 @@ func (tsc *TaskStatusCount) IncrementStatus(status string, statusDetails apimode
 	case evergreen.TaskInactive:
 		tsc.Inactive++
 	}
+}
+
+const jqlBFQuery = "(project in (%v)) and ( %v ) order by updatedDate desc"
+
+// Generates a jira JQL string from the task
+// When we search in jira for a task we search in the specified JIRA project
+// If there are any test results, then we only search by test file
+// name of all of the failed tests.
+// Otherwise we search by the task name.
+func (t *Task) GetJQL(searchProjects []string) string {
+	var jqlParts []string
+	var jqlClause string
+	for _, testResult := range t.LocalTestResults {
+		if testResult.Status == evergreen.TestFailedStatus {
+			fileParts := eitherSlash.Split(testResult.TestFile, -1)
+			jqlParts = append(jqlParts, fmt.Sprintf("text~\"%v\"", fileParts[len(fileParts)-1]))
+		}
+	}
+	if jqlParts != nil {
+		jqlClause = strings.Join(jqlParts, " or ")
+	} else {
+		jqlClause = fmt.Sprintf("text~\"%v\"", t.DisplayName)
+	}
+
+	return fmt.Sprintf(jqlBFQuery, strings.Join(searchProjects, ", "), jqlClause)
 }
