+++ plugin/builtin/manifest/manifest_plugin.go
@@ -2,14 +2,10 @@ package manifest
 
 import (
 	"encoding/json"
-	"fmt"
 	"net/http"
 
 	"github.com/evergreen-ci/evergreen/model/manifest"
-	"github.com/evergreen-ci/evergreen/model/version"
 	"github.com/evergreen-ci/evergreen/plugin"
-	"github.com/evergreen-ci/evergreen/util"
-	"github.com/gorilla/mux"
 )
 
 const (
@@ -72,36 +68,5 @@ func (m *ManifestPlugin) GetPanelConfig() (*plugin.PanelConfig, error) {
 }
 
 func (m *ManifestPlugin) GetUIHandler() http.Handler {
-	r := mux.NewRouter()
-	r.Path("/get/{project_id}/{revision}").HandlerFunc(m.GetManifest)
-	return r
-}
-
-func (m *ManifestPlugin) GetManifest(w http.ResponseWriter, r *http.Request) {
-	project := mux.Vars(r)["project_id"]
-	revision := mux.Vars(r)["revision"]
-
-	version, err := version.FindOne(version.ByProjectIdAndRevision(project, revision))
-	if err != nil {
-		http.Error(w, fmt.Sprintf("error getting version for project %v with revision %v: %v",
-			project, revision, err), http.StatusBadRequest)
-		return
-	}
-	if version == nil {
-		http.Error(w, fmt.Sprintf("version not found for project %v, with revision %v", project, revision),
-			http.StatusNotFound)
-		return
-	}
-
-	foundManifest, err := manifest.FindOne(manifest.ById(version.Id))
-	if err != nil {
-		http.Error(w, fmt.Sprintf("error getting manifest with version id %v: %v",
-			version.Id, err), http.StatusBadRequest)
-		return
-	}
-	if foundManifest == nil {
-		http.Error(w, fmt.Sprintf("manifest not found for version %v", version.Id), http.StatusNotFound)
-		return
-	}
-	util.WriteJSON(w, http.StatusOK, foundManifest)
+	return nil
 }
