+++ service/ui_plugin_build_baron_tickets_test.go
@@ -1,4 +1,4 @@
-package buildbaron
+package service
 
 import (
 	"testing"
