+++ service/ui_plugin_perf_json.go
@@ -1,4 +1,4 @@
-package taskdata
+package service
 
 import (
 	"net/http"
@@ -10,9 +10,53 @@ import (
 	mgo "gopkg.in/mgo.v2"
 )
 
+func perfGetTasksForVersion(w http.ResponseWriter, r *http.Request) {
+	jsonForTasks, err := model.PerfGetTasksForVersion(mux.Vars(r)["version_id"], mux.Vars(r)["name"])
+	if jsonForTasks == nil {
+		http.Error(w, "{}", http.StatusNotFound)
+		return
+	}
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+	util.WriteJSON(w, http.StatusOK, jsonForTasks)
+}
+
+func perfGetTasksForLatestVersion(w http.ResponseWriter, r *http.Request) {
+	project := mux.Vars(r)["project_id"]
+	name := mux.Vars(r)["name"]
+	skip, err := util.GetIntValue(r, "skip", 0)
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusBadRequest)
+		return
+	}
+	if skip < 0 {
+		http.Error(w, "negative skip", http.StatusBadRequest)
+		return
+	}
+
+	data, err := model.PerfGetTasksForLatestVersion(project, name, skip)
+	if err != nil {
+		if err == mgo.ErrNotFound {
+			http.Error(w, "{}", http.StatusNotFound)
+			return
+		}
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+
+	util.WriteJSON(w, http.StatusOK, data)
+}
+
+// getVersion returns a StatusOK if the route is hit
+func perfGetVersion(w http.ResponseWriter, r *http.Request) {
+	util.WriteJSON(w, http.StatusOK, "1")
+}
+
 // FIXME Returns random object that matches conditions
 //       Given set of parameters is not specific enough
-func uiGetCommit(w http.ResponseWriter, r *http.Request) {
+func perfGetCommit(w http.ResponseWriter, r *http.Request) {
 	projectId := mux.Vars(r)["project_id"]
 	revision := mux.Vars(r)["revision"]
 	variant := mux.Vars(r)["variant"]
@@ -34,7 +78,7 @@ func uiGetCommit(w http.ResponseWriter, r *http.Request) {
 	util.WriteJSON(w, http.StatusOK, jsonForTask)
 }
 
-func uiGetTaskHistory(w http.ResponseWriter, r *http.Request) {
+func perfGetTaskHistory(w http.ResponseWriter, r *http.Request) {
 	t, err := task.FindOne(task.ById(mux.Vars(r)["task_id"]))
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
@@ -54,7 +98,7 @@ func uiGetTaskHistory(w http.ResponseWriter, r *http.Request) {
 }
 
 // uiGetTaskById sends back a JSONTask with the corresponding task id.
-func uiGetTaskById(w http.ResponseWriter, r *http.Request) {
+func perfGetTaskById(w http.ResponseWriter, r *http.Request) {
 	taskId := mux.Vars(r)["task_id"]
 	name := mux.Vars(r)["name"]
 	jsonForTask, err := model.GetTaskJSONById(taskId, name)
@@ -69,7 +113,7 @@ func uiGetTaskById(w http.ResponseWriter, r *http.Request) {
 	util.WriteJSON(w, http.StatusOK, jsonForTask)
 }
 
-func uiGetTags(w http.ResponseWriter, r *http.Request) {
+func perfGetTags(w http.ResponseWriter, r *http.Request) {
 	taskId := mux.Vars(r)["task_id"]
 	tags, err := model.GetTaskJSONTags(taskId)
 	if err != nil {
@@ -80,7 +124,7 @@ func uiGetTags(w http.ResponseWriter, r *http.Request) {
 }
 
 // handleTaskTags will update the TaskJSON's tags depending on the request.
-func uiHandleTaskTag(w http.ResponseWriter, r *http.Request) {
+func perfHandleTaskTag(w http.ResponseWriter, r *http.Request) {
 	taskId := mux.Vars(r)["task_id"]
 	name := mux.Vars(r)["name"]
 
@@ -117,7 +161,7 @@ func uiHandleTaskTag(w http.ResponseWriter, r *http.Request) {
 // Wrapper `GetDistinctTagNames` model function
 // requires HTTP GET param `project_id`
 // Return distinct list of tags for given `project_id`
-func uiGetProjectTags(w http.ResponseWriter, r *http.Request) {
+func perfGetProjectTags(w http.ResponseWriter, r *http.Request) {
 	projectIdVals, ok := r.URL.Query()["project_id"]
 
 	if !ok || len(projectIdVals) != 1 {
@@ -134,7 +178,7 @@ func uiGetProjectTags(w http.ResponseWriter, r *http.Request) {
 	util.WriteJSON(w, http.StatusOK, data)
 }
 
-func uiGetTaskJSONByTag(w http.ResponseWriter, r *http.Request) {
+func perfGetTaskJSONByTag(w http.ResponseWriter, r *http.Request) {
 	projectId := mux.Vars(r)["project_id"]
 	tag := mux.Vars(r)["tag"]
 	variant := mux.Vars(r)["variant"]
