+++ service/ui_plugin_perfdash_test.go
@@ -1,4 +1,4 @@
-package perfdash
+package service
 
 import (
 	"testing"
